pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "flask>=3.1.1",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "python-telegram-bot[all]==21.5",
]
